Rename all the common services
